refactor(storage): extract schema v4_tests to sibling file
Pure verbatim relocation of the `#[cfg(test)] mod v4_tests` block from
`primer-storage/src/schema.rs` into a new sibling `src/schema/v4_tests.rs`.
No behaviour change, no incidental fixes.

schema.rs drops 959 → 623 lines. The parent now holds only production
schema/migration code plus a 2-line `#[cfg(test)] mod v4_tests;`
declaration. `schema` is a non-root module (`mod schema;` in lib.rs), so
the submodule resolves to `src/schema/v4_tests.rs` (new `schema/` dir).

The test module is a descendant, so `use super::*` reaches every private
parent item unchanged — zero visibility churn. The ninth test-module
extraction in the owner-approved low-risk oversized-file series
(#268–#275).

Verification:
- cargo test -p primer-storage --lib: 154 passed (baseline 154)
- cargo test --workspace: 51 ok blocks, 0 failures
- cargo clippy -p primer-storage --all-targets -D warnings: clean
- cargo fmt --all --check: clean
